src/eevee/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "eevee"
version = "0.0.1"
description = "eevee analysis program for aixnr/linken pipeline"
requires-python = ">=3.9"
dependencies = [
    "pandas==2.1.1",
    "matplotlib==3.8.0",
    "numpy==1.26.0",
    "pydantic==2.3.0",
    "biopython==1.81",
    "pyfaidx==0.7.2.2"
]

[tool.setuptools]
packages = ["coverage", "substitution", "eevee"]

[project.scripts]
eevee = "eevee.eevee:main"
